PMIx updates
have split up the publish phase for heaps so the base and size
are published separately.   The bundled version was causing
hangs in new PMIx/PRRTE (which is probably something in PRRTE
but we can revert if asked to test).

Fix PMIX type (was working fine but PMIX_INT is "closer"
to the intent).